Refactor control tests into separate file, move Ar code into ar.go. Add GetFilename function
